backend/modules/chatbot/python_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic>=2.0

langchain==0.1.17
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
scikit-learn

pymongo
python-dotenv

# PDF / Web loaders
pypdf
pymupdf
requests
beautifulsoup4
